Backend/requirements.txt
asgiref==3.4.1
Django==3.2.12
psycopg2-binary>=2.8,<2.9
django-crispy-forms==1.13.0
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
djangorestframework==3.12.2
django-cors-headers==3.1.1
dj-rest-auth
django-allauth
django-rest-auth
drf-yasg